Update `src/handlers/web3` to use TypeScript and other impacted files.
